Do not copy the Excel file at the second run if the file is already at the destination
The goal is to let the user fix the file if it is corrupted so that the seeder can correctly generate the seeding files

+++ src/SeederGenerator/Program.cs
@@ -83,6 +83,9 @@ namespace SeederGenerator
                 var sourceFile = Path.Combine(sourcePath, fileName);
                 var destFile = Path.Combine(targetPath, fileName);
 
+                if (File.Exists(destFile))
+                    return;
+
                 File.Copy(sourceFile, destFile, true);
             });
